refactor: (#78) property 하위 oauth 디렉토리를 생성하여 oauth property를 관리한다
