+++ src/main/java/spring/backend/core/configuration/property/oauth/NaverOAuthProperty.java
@@ -1,10 +1,10 @@
-package spring.backend.core.configuration.property;
+package spring.backend.core.configuration.property.oauth;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
-import spring.backend.core.configuration.property.shared.BaseOAuthProperty;
+import spring.backend.core.configuration.property.oauth.shared.BaseOAuthProperty;
 
 @Component
 @Getter
